+++ Changelog.md
@@ -1,4 +1,7 @@
 Change log:
+* 2022-09-22 - U33.2 - 4+2 new pets
+  * Night Sky Carban, Lantern Gourd-lurker, Loyal Turtle, Friendly Heron
+  * Future: Friendly Stork, Friendly Egret
 * 2022-08-22 - U33.1.1 - 1 new pet from Season 3 Reward Track:
   * Black Bat
 * 2022-07-25 - U33.1 - 5 new + 3 newly available: